+++ CHANGELOG.md
@@ -67,6 +67,7 @@ v1.8.0 [unreleased]
 - [#352](https://github.com/latera/homs/pull/352) Wrap task list page together.
 - [#351](https://github.com/latera/homs/pull/351) Improve claim button styles on widget form.
 - [#354](https://github.com/latera/homs/pull/354) Wrap task short pop up list page together.
+- [#353](https://github.com/latera/homs/pull/353) Integrate new claiming components.
 
 v1.7.8 [2019-10-03]
 -------------------